addressing comments and putting method call in separate thread
